requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2; python_full_version >= '3.5.3'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
bidict==0.20.0
chardet==3.0.4
discord.py==1.4.1; python_full_version >= '3.5.3'
discord==1.0.1
humanreadable==0.1.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
mbstrdecoder==1.0.0; python_version >= '3.5'
multidict==4.7.6; python_version >= '3.5'
python-decouple==3.3
typepy==1.1.1; python_version >= '3.5'
yarl==1.5.1
